ResNet50_inference/requirements.txt
aiocontextvars==0.2.2
anyio==3.5.0
asgiref==3.4.1
attrs==21.4.0
braceexpand==0.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
checksumdir==1.2.0
click==8.0.4
cloudpickle==2.0.0
configparser==5.2.0
contextlib2==21.6.0
contextvars==2.4
crc32c==2.2.post0
dataclasses==0.8
docker-pycreds==0.4.0
execnet==1.9.0
fastapi==0.75.2
gitdb==4.0.9
GitPython==3.1.18
h11==0.13.0
horovod==0.24.2
idna==3.3
immutables==0.17
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
loguru==0.6.0
numpy==1.19.5
packaging==21.3
pathtools==0.1.2
Pillow==8.4.0
pluggy==1.0.0
promise==2.3
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
pytest==6.2.5
pytest-forked==1.3.0
pytest-pythonpath==0.7.3
pytest-xdist==2.1.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
sentry-sdk==1.5.8
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.17.1
subprocess32==3.5.4
tfrecord==1.14.1
timm==0.4.12
toml==0.10.2
torchvision==0.11.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.16.0
wandb==0.12.1
webdataset==0.1.54
zipp==3.6.0
